requirements.txt
streamlit>=1.41.1
pdf2image>=1.16.3
Pillow>=11.0.0
markdown>=3.7
pytesseract>=0.3.10
arabic-reshaper>=3.0.0
python-bidi>=0.6.3
requests>=2.32.3
python-docx>=1.1.2
PyPDF2>=3.0.1
textblob>=0.17.1
langdetect>=1.0.9
numpy>=1.24.0
mdutils>=1.6.0
scikit-learn>=1.3.0
pandas>=2.1.0
plotly>=5.18.0
tabula-py>=2.9.0
opencv-python>=4.8.0
deskew>=0.10.30
scikit-image>=0.21.0